Get organization sidebar data from a new hook, removing the complex useEffect in UserProfileSidebar
